// back-end/src/types/labelData.ts
export type LabelData = {
  readonly timestamp: number;
  readonly uuid: string;
  readonly cameraIndex: number;
  readonly cameraType: string;
  readonly width: number;
  readonly height: number;
  readonly session: string;
  readonly store_id: number;
  readonly x: number;
  readonly y: number;
  readonly w: number;
  readonly h: number;
  readonly source_uuid: string;
};